requirements.txt
annotated-types==0.7.0
python-dotenv==1.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.4
curl_cffi==0.13.0
db-dtypes==1.5.0
defusedxml==0.7.1
frozendict==2.4.7
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
multitasking==0.0.12
numpy==2.4.2
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
pandas-gbq==0.33.0
platformdirs==4.5.1
proto-plus==1.27.1
protobuf==5.29.5
psutil==7.2.2
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydata-google-auth==1.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.10.2
six==1.17.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
yfinance==1.1.0
